code-ta/requirements.txt
django-filter==23.3
Django==4.0.5
djangorestframework==3.13.1
drf-spectacular==0.22.1
docker==6.1.3
Markdown==3.4.4